util: Set up line buffering for stderr in setProgName()
Commit 0c06d25e278412d9280c122005ddf0d097f89078 upstream.

If two or more processes share stderr, output can appear mixed up, as
all of the print and hexdump functions use more than one output call per
invocation and stderr is unbuffered by default.

Enable line-buffering for stderr to reduce this effect. This will not
eliminate the problem completely, as there appears to be no guarantee
that buffers are flushed atomically across process boundaries. Still,
the resulting output looks much better.

Perform the setup in setProgName(), as this is meant to be called before
any output function.