util: Fully buffer stderr, flush before unlocking
Commit 10fb5246b863d2a8ea43a8d5e2dd24f82cc98607 upstream.

In Commit 322f51d95bcd84a20b4bd62026f5cf17506a285d we enabled line
buffering for stderr. The problem with this is that the output style is
reset to STYLE_NORMAL _after_ the newline character is printed, i.e.
after the buffer is flushed. Therefore, effectively the output style of
the terminal is not reset between two invocations of the print
functions. If another program prints to the terminal in the meantime,
its output will appear in the same style as initially set up by the
print function, instead of the default style of the terminal. This is
not what we want.

To fix this, call fflush() in unlockFile() before actually unlocking the
file. As we now flush the buffer explicitly, there is no point in using
line buffering anymore, so switch to full buffering.