CMakeLists.txt: Promote some warnings to errors
Again and again I see students ignoring warnings about missing function
declarations (usually due to missing include directives) and missing
return statements. This is a frequent cause for bugs that are hard to
find in the code, as the lines where the functions are called look
correct.

The "implicit int" rule and implicit function declarations are remnants
from the C stone age. New code should not rely on that.

Returning without a value from a non-void function causes undefined
behavior, if the caller expects a return value, so it really should be
avoided. Also throw an error for these cases.