gitignore: Ignore build artifacts outside build dir, too
In case the project is built in the root directory of the project and
not (as intended) in the build/ subdirectory, the build process
creates several files and directories in the root directory. Ignore
these files in git.

This does not mean that the project should be built there instead of in
the build/ subdirectory, it just prevents build artifacts from an
accidental build outside of build/ to creep into the git repository.

Thanks to Paul V. Spieß for the idea and the file list.

Signed-off-by: Stefan Gast <stefan.gast@rwu.de>