network: Add TODO about message structs
Add a TODO comment telling the students to define the message structures
for the network protocol when they are implementing the login after the
first exercise sheet. The layout of the structures can be derived from a
document which can be downloaded from Moodle.

Thanks to Gerald Lautenschlager <gerald.lautenschlager@rwu.de> for this
suggestion.