CMakeLists.txt: Specify language in PROJECT statement
If the language is not specified, the default is C and C++. As this
project is C only, add the correct specification, so that CMake does not
check for a C++ compiler, which we do not need for this project anyway.

Thanks to Julian Podiebrad <julian.podiebrad@hs-weingarten.de> for the
hint. I owe you a coffee now. :)